new - 필요없는 클래스 제거 및 신규 클래스 추가
