+++ src/main/java/step2/domain/LottoNumber.java
@@ -1,0 +1,2 @@
+package step2.domain;public class LottoNumber {
+}
