+++ src/test/java/step2/domain/LottoNumberTest.java
@@ -1,0 +1,2 @@
+package step2.domain;public class LottoNumberTest {
+}
